+++ modules/main/src/main/java/com/nisovin/shopkeepers/util/java/MathUtils.java
@@ -8,10 +8,16 @@ import java.util.concurrent.ThreadLocalRandom;
 public final class MathUtils {
 
 	/**
-	 * The default maximum difference between two floating point numbers that are still considered
-	 * equal in a fuzzy comparison.
+	 * The default maximum difference between two double numbers that are still considered equal in
+	 * a fuzzy comparison.
 	 */
 	public static final double EPSILON = 0.00001D;
+
+	/**
+	 * The default maximum difference between two float numbers that are still considered equal in a
+	 * fuzzy comparison.
+	 */
+	public static final float FLOAT_EPSILON = (float) EPSILON;
 
 	/**
 	 * Checks if the given values are roughly equal, using {@link #EPSILON} as the maximum
@@ -49,6 +55,45 @@ public final class MathUtils {
 	 */
 	public static boolean fuzzyEquals(double a, double b, double tolerance) {
 		return Double.compare(a, b) == 0 // Accounts for NaN and infinities
+				|| Math.abs(a - b) <= tolerance;
+	}
+
+	/**
+	 * Checks if the given values are roughly equal, using {@link #FLOAT_EPSILON} as the maximum
+	 * tolerance.
+	 * 
+	 * @param a
+	 *            the first value
+	 * @param b
+	 *            the second value
+	 * @return <code>true</code> if the values are considered equal
+	 * @see #fuzzyEquals(float, float, float)
+	 */
+	public static boolean fuzzyEquals(float a, float b) {
+		return fuzzyEquals(a, b, FLOAT_EPSILON);
+	}
+
+	/**
+	 * Checks if the given values are roughly equal within the given {@code tolerance}.
+	 * <p>
+	 * Notes on the comparison of special values:
+	 * <ul>
+	 * <li>All {@link Float#NaN} values are considered equal.
+	 * <li>{@link Float#POSITIVE_INFINITY} and {@link Float#NEGATIVE_INFINITY} are each considered
+	 * equal to themselves.
+	 * <li>Positive and negative zero are considered equal.
+	 * </ul>
+	 * 
+	 * @param a
+	 *            the first value
+	 * @param b
+	 *            the second value
+	 * @param tolerance
+	 *            the maximum difference between both values to still be considered equal
+	 * @return <code>true</code> if the values are considered equal
+	 */
+	public static boolean fuzzyEquals(float a, float b, float tolerance) {
+		return Float.compare(a, b) == 0 // Accounts for NaN and infinities
 				|| Math.abs(a - b) <= tolerance;
 	}