+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 Date format: (YYYY-MM-DD)  
 
-## v2.23.5 (TBA)
+## v2.23.6 (TBA)
+### Supported MC versions: 1.21.5, 1.21.4, 1.21.3, 1.21.1, 1.21, 1.20.6
+
+
+## v2.23.5 (2025-04-23)
 ### Supported MC versions: 1.21.5, 1.21.4, 1.21.3, 1.21.1, 1.21, 1.20.6
 
 * If available, use the item's custom display name in trade notification messages.